+++ samples/src/simbio/se/sau/sample/MainActivity.java
@@ -3,6 +3,7 @@ package simbio.se.sau.sample;
 import static simbio.se.sau.iterable.Range.range;
 import simbio.se.sau.device.DeviceInformationsManager;
 import simbio.se.sau.log.SimbiLog;
+import simbio.se.sau.sample.documents.ActivityDocumentUtils;
 import simbio.se.sau.share.SimbiShare;
 import simbio.se.sau.view.ClipboardTextView;
 import simbio.se.sau.view.ClipboardTextView.CopiedInterface;
@@ -94,6 +95,10 @@ public class MainActivity extends Activity implements AnimationListener, IRangeS
 
 	public void openSqlActivity(View view) {
 		startActivity(new Intent(getApplicationContext(), SqlExampleActivity.class));
+	}
+
+	public void openDocActivity(View view) {
+		startActivity(new Intent(getApplicationContext(), ActivityDocumentUtils.class));
 	}
 
 	public void showToaster(View view) {